Move pkg/response to kubeops
